Override default endpoint during compilation
Prior to this commit, projects that depended on `forcex` would generate
Salesforce object modules during compilation by using the production URL
to enumerate all available objects to be rendered.  By hardcoding this,
it was not possible to generate objects from a sandbox region.

This commit introduces an additional step to select the URL to be used
when calling the SFDC endpoint during the app compilation step. With
this change, the compilation step will first use the endpoint specified
in the app configuration, defaulting to the production URL if not
otherwise provided.